Moved openinvscode code to its own file
